+++ stripe/src/main/java/com/stripe/android/StripeRepository.kt
@@ -41,7 +41,7 @@ internal interface StripeRepository {
         APIConnectionException::class,
         APIException::class
     )
-    fun confirmPaymentIntent(
+    suspend fun confirmPaymentIntent(
         confirmPaymentIntentParams: ConfirmPaymentIntentParams,
         options: ApiRequest.Options,
         expandFields: List<String> = emptyList()
@@ -77,7 +77,7 @@ internal interface StripeRepository {
         APIConnectionException::class,
         APIException::class
     )
-    fun confirmSetupIntent(
+    suspend fun confirmSetupIntent(
         confirmSetupIntentParams: ConfirmSetupIntentParams,
         options: ApiRequest.Options,
         expandFields: List<String> = emptyList()
@@ -184,7 +184,7 @@ internal interface StripeRepository {
         APIException::class,
         CardException::class
     )
-    fun addCustomerSource(
+    suspend fun addCustomerSource(
         customerId: String,
         publishableKey: String,
         productUsageTokens: Set<String>,
@@ -200,7 +200,7 @@ internal interface StripeRepository {
         APIException::class,
         CardException::class
     )
-    fun deleteCustomerSource(
+    suspend fun deleteCustomerSource(
         customerId: String,
         publishableKey: String,
         productUsageTokens: Set<String>,
@@ -215,7 +215,7 @@ internal interface StripeRepository {
         APIException::class,
         CardException::class
     )
-    fun attachPaymentMethod(
+    suspend fun attachPaymentMethod(
         customerId: String,
         publishableKey: String,
         productUsageTokens: Set<String>,
@@ -230,7 +230,7 @@ internal interface StripeRepository {
         APIException::class,
         CardException::class
     )
-    fun detachPaymentMethod(
+    suspend fun detachPaymentMethod(
         publishableKey: String,
         productUsageTokens: Set<String>,
         paymentMethodId: String,
@@ -244,7 +244,7 @@ internal interface StripeRepository {
         APIException::class,
         CardException::class
     )
-    fun getPaymentMethods(
+    suspend fun getPaymentMethods(
         listPaymentMethodsParams: ListPaymentMethodsParams,
         publishableKey: String,
         productUsageTokens: Set<String>,
@@ -258,7 +258,7 @@ internal interface StripeRepository {
         APIException::class,
         CardException::class
     )
-    fun setDefaultCustomerSource(
+    suspend fun setDefaultCustomerSource(
         customerId: String,
         publishableKey: String,
         productUsageTokens: Set<String>,
@@ -274,7 +274,7 @@ internal interface StripeRepository {
         APIException::class,
         CardException::class
     )
-    fun setCustomerShippingInfo(
+    suspend fun setCustomerShippingInfo(
         customerId: String,
         publishableKey: String,
         productUsageTokens: Set<String>,
@@ -289,7 +289,7 @@ internal interface StripeRepository {
         APIException::class,
         CardException::class
     )
-    fun retrieveCustomer(
+    suspend fun retrieveCustomer(
         customerId: String,
         productUsageTokens: Set<String>,
         requestOptions: ApiRequest.Options
@@ -340,7 +340,7 @@ internal interface StripeRepository {
         requestOptions: ApiRequest.Options
     ): Complete3ds2Result
 
-    fun createFile(
+    suspend fun createFile(
         fileParams: StripeFileParams,
         requestOptions: ApiRequest.Options
     ): StripeFile