+++ stripe/src/main/java/com/stripe/android/view/CvcEditText.kt
@@ -2,12 +2,11 @@ package com.stripe.android.view
 
 import android.content.Context
 import android.os.Build
-import android.text.Editable
 import android.text.InputFilter
 import android.text.InputType
-import android.text.method.DigitsKeyListener
 import android.util.AttributeSet
 import android.view.View
+import androidx.core.widget.doAfterTextChanged
 import com.google.android.material.textfield.TextInputLayout
 import com.stripe.android.R
 import com.stripe.android.cards.Cvc
@@ -53,23 +52,18 @@ open class CvcEditText @JvmOverloads constructor(
         maxLines = 1
         filters = createFilters(CardBrand.Unknown)
 
-        inputType = InputType.TYPE_NUMBER_VARIATION_PASSWORD
-        keyListener = DigitsKeyListener.getInstance(false, true)
+        inputType = InputType.TYPE_CLASS_NUMBER
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             setAutofillHints(View.AUTOFILL_HINT_CREDIT_CARD_SECURITY_CODE)
         }
 
-        addTextChangedListener(
-            object : StripeTextWatcher() {
-                override fun afterTextChanged(s: Editable?) {
-                    shouldShowError = false
-                    if (cardBrand.isMaxCvc(unvalidatedCvc.normalized)) {
-                        completionCallback()
-                    }
-                }
+        doAfterTextChanged {
+            shouldShowError = false
+            if (cardBrand.isMaxCvc(unvalidatedCvc.normalized)) {
+                completionCallback()
             }
-        )
+        }
     }
 
     override val accessibilityText: String?