+++ stripe/src/test/java/com/stripe/android/StripeApiRepositoryTest.kt
@@ -372,7 +372,7 @@ internal class StripeApiRepositoryTest {
     }
 
     @Test
-    fun confirmPaymentIntent_withSourceData_canSuccessfulConfirm() {
+    fun confirmPaymentIntent_withSourceData_canSuccessfulConfirm() = testDispatcher.runBlockingTest {
         // put a private key here to simulate the backend
         val clientSecret = "pi_12345_secret_fake"
 
@@ -413,7 +413,7 @@ internal class StripeApiRepositoryTest {
     }
 
     @Ignore("requires a secret key")
-    fun disabled_confirmPaymentIntent_withSourceId_canSuccessfulConfirm() {
+    fun disabled_confirmPaymentIntent_withSourceId_canSuccessfulConfirm() = testDispatcher.runBlockingTest {
         val clientSecret = "temporarily put a private key here simulate the backend"
         val publishableKey = "put a public key that matches the private key here"
         val sourceId = "id of the source created on the backend"
@@ -471,7 +471,7 @@ internal class StripeApiRepositoryTest {
     }
 
     @Test
-    fun getPaymentMethods_whenPopulated_returnsExpectedList() {
+    fun getPaymentMethods_whenPopulated_returnsExpectedList() = testDispatcher.runBlockingTest {
         val responseBody =
             """
             {
@@ -648,7 +648,7 @@ internal class StripeApiRepositoryTest {
     }
 
     @Test
-    fun getPaymentMethods_whenNotPopulated_returnsEmptydList() {
+    fun getPaymentMethods_whenNotPopulated_returnsEmptyList() = testDispatcher.runBlockingTest {
         val responseBody =
             """
             {
@@ -759,7 +759,7 @@ internal class StripeApiRepositoryTest {
     }
 
     @Test
-    fun createFile_shouldFireExpectedRequests() {
+    fun createFile_shouldFireExpectedRequests() = testDispatcher.runBlockingTest {
         val stripeRepository = create()
 
         stripeRepository.createFile(
@@ -772,7 +772,8 @@ internal class StripeApiRepositoryTest {
 
         verify(stripeApiRequestExecutor, never()).execute(any<ApiRequest>())
         verify(stripeApiRequestExecutor).execute(fileUploadRequestArgumentCaptor.capture())
-        assertNotNull(fileUploadRequestArgumentCaptor.firstValue)
+        assertThat(fileUploadRequestArgumentCaptor.firstValue)
+            .isNotNull()
 
         verifyAnalyticsRequest(AnalyticsEvent.FileCreate)
     }